Address few remaining comments
