+++ test/support/utils.h
@@ -29,6 +29,8 @@
 #include <cstring>
 #include <iostream>
 #include <iterator>
+#include <complex>
+#include <type_traits>
 #include <memory>
 #include <sstream>
 #include <vector>
@@ -691,6 +693,46 @@ struct can_use_default_less_operator : ::std::false_type
 template <typename T>
 struct can_use_default_less_operator<T, decltype(::std::declval<T>() < ::std::declval<T>())> : ::std::true_type
 {
+};
+
+// An arbitrary binary predicate to simulate a predicate the user providing
+// a custom predicate.
+template <typename _Tp>
+struct UserBinaryPredicate
+{
+    bool
+    operator()(const _Tp& __x, const _Tp& __y) const
+    {
+        using KeyT = ::std::decay_t<_Tp>;
+        return __y != KeyT(1);
+    }
+};
+
+template <typename _Tp>
+struct MaxFunctor
+{
+    _Tp
+    operator()(const _Tp& __x, const _Tp& __y) const
+    {
+        return (__x < __y) ? __y : __x;
+    }
+};
+
+// TODO: Investigate why we cannot call ::std::abs on complex
+// types with the CUDA backend.
+template <typename _Tp>
+struct MaxFunctor<::std::complex<_Tp>>
+{
+    auto
+    complex_abs(const ::std::complex<_Tp>& __x) const
+    {
+        return ::std::sqrt(__x.real() * __x.real() + __x.imag() * __x.imag());
+    }
+    ::std::complex<_Tp>
+    operator()(const ::std::complex<_Tp>& __x, const ::std::complex<_Tp>& __y) const
+    {
+        return (complex_abs(__x) < complex_abs(__y)) ? __y : __x;
+    }
 };
 
 } /* namespace TestUtils */
